feat(WaitingSavePolicy): 저장 정책 관련 역할 분리
